wip: select server
