+++ packages/api/src/index.ts
@@ -1,5 +1,5 @@
 export { authApi, authQueryKeys } from './auth'
-export { API, apiIsInitialized, checkUrl, initializeApi, isUrl } from './axios'
+export { API, apiIsInitialized, checkUrl, formatServiceURL, initializeApi, isUrl } from './axios'
 export { bookClubApi, bookClubQueryKeys } from './bookClub'
 export { emailerApi, emailerQueryKeys } from './emailer'
 export { epubApi, epubQueryKeys, getEpubResource, updateEpubProgress } from './epub'
