+++ packages/browser/src/components/savedServer/configuredServers/ConfiguredServer.tsx
@@ -1,8 +1,7 @@
 import { IconButton, Text, ToolTip } from '@stump/components'
 import { useLocaleContext } from '@stump/i18n'
 import { SavedServer } from '@stump/types'
-import { BadgeCheck, Power, Settings2, Trash2 } from 'lucide-react'
-import React from 'react'
+import { BadgeCheck, Power, Settings2, Trash2, WifiOff } from 'lucide-react'
 
 type Props = {
 	/**
@@ -25,18 +24,39 @@ type Props = {
 	 * A callback to trigger the switch confirmation modal to render
 	 */
 	onSwitch: () => void
+	/**
+	 * Whether the server is reachable
+	 */
+	isReachable?: boolean
 }
 
-export default function ConfiguredServer({ server, isActive, onEdit, onDelete, onSwitch }: Props) {
+export default function ConfiguredServer({
+	server,
+	isActive,
+	onEdit,
+	onDelete,
+	onSwitch,
+	isReachable,
+}: Props) {
 	const { t } = useLocaleContext()
 
 	return (
-		<div className="group flex items-center justify-between">
+		<div className="group flex items-center justify-between p-4">
 			<div className="flex flex-col">
 				<span className="flex items-center space-x-2">
 					<Text>{server.name}</Text>
 					{isActive && (
-						<BadgeCheck className="h-4 w-4 text-fill-success text-opacity-75" strokeWidth={0.95} />
+						<ToolTip content={t(getKey('activeServer.tooltip'))} align="center">
+							<BadgeCheck
+								className="h-4 w-4 text-fill-success text-opacity-75"
+								strokeWidth={0.95}
+							/>
+						</ToolTip>
+					)}
+					{isReachable === false && (
+						<ToolTip content={t(getKey('unreachableServer.tooltip'))} align="center">
+							<WifiOff className="h-4 w-4 text-fill-danger text-opacity-75" />
+						</ToolTip>
 					)}
 				</span>
 				<Text variant="muted" size="sm">
