+++ packages/browser/src/components/savedServer/configuredServers/DeleteServerConfirmation.tsx
