+++ packages/browser/src/components/savedServer/configuredServers/ResetConfigurationConfirmation.tsx
