+++ packages/browser/src/components/savedServer/configuredServers/ResetConfiguredServersSection.tsx
