+++ packages/browser/src/components/savedServer/configuredServers/index.ts
@@ -1,0 +1,1 @@
+export { default as ConfiguredServersList } from './ConfiguredServersList'
