+++ packages/browser/src/components/savedServer/index.ts
@@ -1,1 +1,2 @@
+export { ConfiguredServersList } from './configuredServers'
 export * from './createOrUpdate'
